Factor out resolve fns instead of noop
