--- src/noop.js
@@ -1,1 +1,0 @@
-export default () => {};
